% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{PATH_USER_DB}
\alias{PATH_USER_DB}
\title{Global Variable: User database path}
\format{
An object of class \code{fs_path} (inherits from \code{character}) of length 1.
}
\usage{
PATH_USER_DB
}
\description{
Global Variable: User database path
}
\keyword{datasets}
